requirements.txt
# Core
python-dotenv>=1.0.0
pyodbc>=4.0.39

# XML/CFDI
lxml>=4.9.0

# Reportes
openpyxl>=3.1.0

# Matching de productos
rapidfuzz>=3.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
